fix(ci): Correctly pass tags to spring-boot:build-image (#30)
This commit fixes the Docker publishing workflow by correctly passing the
tags to the `spring-boot:build-image` goal. The `tags` output from the
`docker/metadata-action` is a multi-line string, which was causing issues
when passed directly as a Maven property.

This commit now joins the tags with a comma to create a single string
that can be correctly parsed by the `spring-boot-maven-plugin`.

Co-authored-by: Rohit Khatkar <khatkarrohit@gmail.com>